==> src/2_duplicates.ipynb <==
#To be able to install textpack
!conda install -y -c conda-forge sparse_dot_topn

!pip install textpack

import pickle
import duplicates_removal

## Get cleaned emails

with open('dutch_cleaned_emails_id_body_org.pkl', 'rb') as f:
    emails_orgs = pickle.load(f)

## Remove exact duplicates

emails_orgs_no_dup = emails_orgs.drop_duplicates(subset=['only_body'])

## Run textpack

emails_orgs_no_dup = duplicates_removal.duplicates_per_org(emails_orgs_no_dup)

## Remove duplicates

emails_orgs_no_dup['is_different'] = emails_orgs_no_dup.apply(lambda x: duplicates_removal.is_different(x["only_body"], x["duplicate"]), axis = 1)

emails_no_dup = emails_orgs_no_dup[emails_orgs_no_dup['is_different'] == 0]
emails_no_dup = emails_no_dup[["id", "only_body"]]

## Save emails without duplicates at 0.95

with open('no_duplicates_0_95.pkl', 'wb') as f:
     pickle.dump(emails_no_dup, f)
